Unify packed leftover prototypes that disagree on return type.
Fat TUs concatenated files that each declared WaitFrames and other
helpers with void vs s32. After recovered aliases map those names to
one Func_, old_agbcc reports conflicting types. Keep one return type
per name in the file.

+++ games/gs1/src/math/multiply_fixed_8.c
@@ -31,7 +31,7 @@ s16 FixedPoint_ScaleByRatio(s16 arg0, s16 arg1)
 }
 
 /* math/reciprocal_fixed.c */
-s16 FixedPoint_Ratio(s32, s16);
+s32 FixedPoint_Ratio(s32, s16);
 
 s16 FixedPoint_Reciprocal(s16 arg0)
 {
